Move Entities.dumpInfo() methods to Dumper class
